Move migration to oob_migrate
